Eliminate warnnings for VC
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@384 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ OBPRM/Clock_Class.cpp
@@ -173,9 +173,10 @@ GetClock_USEC() {
 
 ///Modified for VC
 #if defined(_WIN32)
-  return u_time*1e6+u_utime*1e3; //here u_utime is millisecond
+  return (int)(u_time*1e6+u_utime*1e3); //here u_utime is millisecond
 #else
-  return u_time*1e6+u_utime;
+  return (int)(u_time*1e6+u_utime);
 #endif
+
 };
